requirements.txt
path.py
Django
south
fabric
django-grappelli
django-extensions
Werkzeug
git+git://github.com/jezdez/django_compressor.git
django-registration
django-simple-captcha
PIL
Pillow